+++ extra-bonus-notifier/calculation.go
@@ -9,12 +9,12 @@ import (
 func CalculateExtraBonus(
 	newsSeen, extraBonusDaysClaimNotAvailable, extraBonusIndex uint16,
 	now, extraBonusLastClaimAvailableAt, miningSessionSoloStartedAt, miningSessionSoloEndedAt *time.Time,
-) uint16 {
+) float64 {
 	const networkDelayDelta = 1.333
 	var (
 		newsSeenBonusValues            = cfg.ExtraBonuses.NewsSeenValues
 		miningStreakValues             = cfg.ExtraBonuses.MiningStreakValues
-		bonusPercentageRemaining       = 100 * (1 + extraBonusDaysClaimNotAvailable)
+		bonusPercentageRemaining       = float64(100 * (1 + extraBonusDaysClaimNotAvailable))
 		miningStreak                   = calculateMiningStreak(now, miningSessionSoloStartedAt, miningSessionSoloEndedAt)
 		flatBonusValue                 = cfg.ExtraBonuses.FlatValues[extraBonusIndex]
 		firstDelayedClaimPenaltyWindow = int64(float64(cfg.ExtraBonuses.DelayedClaimPenaltyWindow.Nanoseconds()) * networkDelayDelta)
@@ -23,7 +23,7 @@ func CalculateExtraBonus(
 		return 0
 	}
 	if delay := now.Sub(*extraBonusLastClaimAvailableAt.Time); delay.Nanoseconds() > firstDelayedClaimPenaltyWindow {
-		bonusPercentageRemaining -= 25 * uint16(delay/cfg.ExtraBonuses.DelayedClaimPenaltyWindow)
+		bonusPercentageRemaining -= 25 * float64(delay/cfg.ExtraBonuses.DelayedClaimPenaltyWindow)
 	}
 	if miningStreak >= uint64(len(miningStreakValues)) {
 		miningStreak = uint64(len(miningStreakValues) - 1)
@@ -32,7 +32,7 @@ func CalculateExtraBonus(
 		newsSeen = uint16(len(newsSeenBonusValues) - 1)
 	}
 
-	return ((flatBonusValue + miningStreakValues[miningStreak] + newsSeenBonusValues[newsSeen]) * bonusPercentageRemaining) / 100
+	return (float64(flatBonusValue+miningStreakValues[miningStreak]+newsSeenBonusValues[newsSeen]) * bonusPercentageRemaining) / 100
 }
 
 func calculateMiningStreak(now, start, end *time.Time) uint64 {
